"moved the polars ls command under the cache module"
