+++ crates/nu_plugin_polars/src/cache/list.rs
